requirements.txt
# VCAI Requirements
# =================
# 
# IMPORTANT: PyTorch must be installed FIRST with CUDA support!
# See docs/SETUP.md for complete setup guide.
#
# Quick setup:
#   1. Install PyTorch with CUDA: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#   2. Then: pip install -r requirements.txt

# ------------------------------------------------------------------------------
# FastAPI and Server
# ------------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# ------------------------------------------------------------------------------
# Database
# ------------------------------------------------------------------------------
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# ------------------------------------------------------------------------------
# Redis (Cache and Session)
# ------------------------------------------------------------------------------
redis>=5.0.1

# ------------------------------------------------------------------------------
# Authentication
# ------------------------------------------------------------------------------
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ------------------------------------------------------------------------------
# Validation
# ------------------------------------------------------------------------------
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# ------------------------------------------------------------------------------
# ML/AI Components
# ------------------------------------------------------------------------------
# NOTE: PyTorch is NOT listed here - install separately with CUDA!
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

faster-whisper>=1.0.1
langgraph>=0.0.26
langchain>=0.1.0
langchain-core>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2

# ------------------------------------------------------------------------------
# Audio Processing
# ------------------------------------------------------------------------------
numpy>=1.26.0
sounddevice>=0.4.6
soundfile>=0.12.1

# ------------------------------------------------------------------------------
# Utilities
# ------------------------------------------------------------------------------
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1

# ------------------------------------------------------------------------------
# Testing
# ------------------------------------------------------------------------------
pytest>=7.4.4
pytest-asyncio>=0.23.3

# ------------------------------------------------------------------------------
# Development
# ------------------------------------------------------------------------------
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0